Stub out the vuex store
Just looking at the shopping cart example github repo for help, I'm not
sure if I really need a module yet or anything like that, but I figure I
might as well just set things up like that for now.

The real thing I'm curious about is where the logic of the game should
be placed? Like, is it all just button presses and fire off events
maybe? I suppose I need a store to keep track of the Deck's current
state so the dealer can count cards and whatnot.